Fix minor formatting and move remaining make test to PottsCellStem
